requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
langchain
langchain-core
langchain-mistralai
httpx